@@ -3,11 +3,11 @@ class ReviewDecorator
 
   attr_reader :review
 
-  def initialize(review)
-    @review = review
+  def initialize(review, opts={ })
+    @review, @reviewee_first_name, @reviewer = review, opts[:reviewee_first_name], opts[:reviewer]
   end
 
-  def decorate
+  def simple_decorate
     {
         :review => review,
         :id => id,
@@ -15,11 +15,14 @@ class ReviewDecorator
     }
   end
 
-  def comment_decorate
+  def decorate
     {
         :comment => comment,
+        :review => review,
         :created_at => created_at,
-        :reviewer => reviewer
+        :reviewee => reviewee_first_name,
+        :reviewer_url_slug => reviewer_url_slug,
+        :reviewer_first_name => reviewer_first_name
     }
   end
 
@@ -40,6 +43,18 @@ class ReviewDecorator
   end
 
   def reviewer
-    User.find(@review[:author_id]).first_name
+    @reviewer ||= User.find(@review[:author_id])
+  end
+
+  def reviewer_first_name
+    reviewer[:first_name] ||= ''
+  end
+
+  def reviewee_first_name
+    @reviewee_first_name ||= User.find(@review[:user_id]).first_name
+  end
+
+  def reviewer_url_slug
+    reviewer[:url_slug] ||= ''
   end
 end
