@@ -14,32 +14,27 @@ import TimePicker from 'material-ui/TimePicker';
 import TextField from 'material-ui/TextField';
 
 import FormData from './utils/FormData';
+import formatLink from './utils/Link';
 
-import Header from './Header';
+import Header from './reusable/Header';
 import ErrorField from './reusable/ErrorField';
-import { postData } from './sendData';
+import { postData } from './utils/sendData';
 
 import './NewAvailability.css';
-import Footer from './Footer';
+import Footer from './reusable/Footer';
+import PageHeader from './reusable/PageHeader';
 
 const nowDate = new Date();
 const DEFAULT_START = _.toString(new Date(nowDate.setHours(0, 0)));
 const DEFAULT_END = _.toString(new Date(nowDate.setHours(23, 59)));
 
 const schema = Joi.object({}).pattern(/[0-9]+/, Joi.object({
-    day: Joi.string().required().options({
+    day: Joi.number().required().options({
       language: {
-        any: {
-          required:
-  <FormattedMessage
-    id='NewAvailability.selectDayError'
-    defaultMessage='Please select a day'
-          />,
-          empty:
-  <FormattedMessage
-    id='NewAvailability.selectDayError'
-    defaultMessage='Please select a day'
-          />
+        number: {
+          base: 'Please select a day',
+          any: 'Please select a day',
+          empty: 'Please select a day'
         }
       }
     }),
@@ -47,11 +42,7 @@ const schema = Joi.object({}).pattern(/[0-9]+/, Joi.object({
       language: {
         date: {
           timestamp: {
-            javascript:
-  <FormattedMessage
-    id='NewAvailability.startTimeBlank'
-    defaultMessage='Please select a start time'
-              />
+            javascript: 'Please select a start time'
           }
         }
       }
@@ -60,54 +51,42 @@ const schema = Joi.object({}).pattern(/[0-9]+/, Joi.object({
       language: {
         date: {
           timestamp: {
-            javascript:
-  <FormattedMessage
-    id='NewAvailability.endTimeBlank'
-    defaultMessage='Please select an end time'
-            />
+            javascript: 'Please select an end time'
           }
         }
       }
     })
 }));
 
-function validateTimes({ values, validateAllValues, changingValues, errors }, callback){
+function validateTimes({ values, errors }, callback){
+  _.each(values, (item, index) => {
 
-  if (validateAllValues || _.some(changingValues, (item) => _.endsWith(item, 'start_time')) || _.some(changingValues, (item) =>_.endsWith(item, 'end_time'))) {
-    _.each(changingValues, (item) => {
-      const index = _.split(item, '.')[0];
+    if (item.start_time && item.end_time) {
+      const endTimeIsBeforeStartTime = item.end_time < item.start_time;
+      const startTimeError =
+        (<FormattedMessage
+          id='NewAvailability.startTimeError'
+          defaultMessage='Please select a start time chronologically before end time'
+        />);
 
-      if (_.endsWith(item, 'start_time') || _.endsWith(item, 'end_time')) {
-        if (values[index]) {
-          if (values[index].start_time && values[index].end_time) {
-            const endTimeIsBeforeStartTime = values[index].end_time < values[index].start_time;
-            const startTimeError =
-              (<FormattedMessage
-                id='NewAvailability.startTimeError'
-                defaultMessage='Please select a start time chronologically before end time'
-              />);
+        const endTimeError =
+          (<FormattedMessage
+            id='NewAvailability.endTimeError'
+            defaultMessage='Please select an end time chronologically after start time'
+        />);
 
-              const endTimeError =
-                (<FormattedMessage
-                  id='NewAvailability.endTimeError'
-                  defaultMessage='Please select an end time chronologically after end time'
-              />);
-
-            if (endTimeIsBeforeStartTime) {
-              if (_.endsWith(item, 'start_time')) { _.set(errors, `${index}.start_time`, startTimeError); }
-              if (_.endsWith(item, 'end_time')) { _.set(errors, `${index}.end_time`, endTimeError); }
-            }
-          }
-
-          if (values[index].start_time && !values[index].end_time) {
-            return values[index].end_time = new Date(nowDate.setHours(23, 59));
-          } else if (values[index].end_time && !values[index].start_time) {
-            return values[index].start_time = new Date(nowDate.setHours(0, 0));
-          }
-        }
+      if (endTimeIsBeforeStartTime) {
+         _.set(errors, `${index}.start_time`, startTimeError);
+         _.set(errors, `${index}.end_time`, endTimeError);
       }
-    });
-  }
+    }
+
+    if (item.start_time && !item.end_time) {
+      return item.end_time = new Date(nowDate.setHours(23, 59));
+    } else if (item.end_time && !item.start_time) {
+      return item.start_time = new Date(nowDate.setHours(0, 0));
+    }
+  });
 
   callback({ values, errors });
 }
@@ -123,7 +102,7 @@ class NewAvailability extends Component {
 
     this.state = {
       numberOfAvailabilities: 1,
-      error: ''
+      error: { }
     };
   }
 
@@ -131,7 +110,10 @@ class NewAvailability extends Component {
     const {
       validateAllHandler,
       currentUser,
-      currentUser: { timezone }
+      currentUser: {
+        timezone,
+        locale
+      }
     } = this.props;
 
     return (
@@ -140,7 +122,6 @@ class NewAvailability extends Component {
 
         <div className='availabilityContainer'>
           { this.renderTitle() }
-          <ErrorField error={ this.state.error } />
 
           <form className='newAvailabilityFormContainer'>
             <div>
@@ -151,7 +132,7 @@ class NewAvailability extends Component {
                 disabled
               />
 
-              <a href="/my_profile" className='slidingLink' >
+              <a href={ formatLink('/my_profile', locale) } className='slidingLink' >
                 <FormattedMessage
                   id='NewAvailability.updateTimezone'
                   defaultMessage='Not your timezone?'
@@ -174,63 +155,83 @@ class NewAvailability extends Component {
             { this.renderAvailabilities() }
           </form>
         </div>
-        <Footer />
+        <Footer className='footerContainerFixed' />
       </div>
     );
   }
 
   renderTitle() {
-    const { match: { params: { sign_up } } } = this.props;
+    const { location: { state }  } = this.props;
 
-    if (sign_up) {
+    if (state && state.signUp) {
       return (
         <h1 className='signUpHeader'>
           <FormattedMessage
             id='signUpHeader'
-            defaultMessage='Join Tutoria community: Step 2/2'
+            defaultMessage='Join Tutoría community: Step 2/2'
           />
         </h1>
+      );
+    } else {
+      return (
+        <PageHeader title={
+          <FormattedMessage
+            id='NewAvailability.header'
+            defaultMessage='Create your availabilities'
+          />
+         }
+        />
       );
     }
   }
 
   handleSubmit() {
-    const { errors, validateAllHandler } = this.props;
+    const { validateAll } = this.props;
 
-    validateAllHandler();
+    validateAll(() => {
+      const { errors } = this.props;
 
-    if (_.size(errors) === 0) {
-      const { availabilities } = this.props;
+      if (_.size(errors) === 0) {
+        const { availabilities, currentUser: { locale } } = this.props;
 
-      const attributes = FormData.from({ availabilities });
-      const requestParams = {
-        url: '/availabilities',
-        attributes,
-        method: 'POST',
-        successCallBack: () => {
-          location.assign('/availabilities');
-        },
-        errorCallBack: (message) => {
-          this.setState({
-            error: message
-          });
-        }
-      };
+        const attributes = FormData.from({ availabilities });
 
-      return postData(requestParams);
-    }
+        const requestParams = {
+          url: '/availabilities',
+          attributes,
+          method: 'POST',
+          successCallBack: () => {
+            location.assign(formatLink('/availabilities', locale));
+          },
+          errorCallBack: (message) => {
+            this.setState({
+              error: message
+            });
+          }
+        };
+
+        return postData(requestParams);
+      }
+    });
+
   }
 
   renderAvailabilities() {
     const { numberOfAvailabilities } = this.state;
-    const { errors, changeHandler, validateHandler, validateAllHandler, days } = this.props;
+    const { changeHandler, validateAllHandler, days } = this.props;
 
     return _.times(numberOfAvailabilities, (index) => {
-      const availability = index;
-      const { availabilities } = this.props;
+      const { availabilities, errors } = this.props;
+      const currentAvailability = availabilities[index] || { };
+      const { day, start_time, end_time } = currentAvailability;
+      const { error } = this.state;
 
       return (
         <div key={ index } className='availabilitiesContainer' >
+          <div className='newAvailabilityErrorField'>
+            <ErrorField error={ error && error[index] } />
+          </div>
+
           <SelectField
             hintText={
               <FormattedMessage
@@ -238,46 +239,47 @@ class NewAvailability extends Component {
                 defaultMessage='Select Day'
               />
             }
-            value={ availabilities[index]? availabilities[index].day: {} }
-            errorText={ _.get(errors, `${availability}.day`) }
-            onChange={ this.changeDayHandler(availability) }
-            onBlur={ validateHandler(`${availability}.day`) }
+            value={ currentAvailability? day: '' }
+            errorText={ _.get(errors, `${index}.day`) }
+            onChange={ changeHandler(`${index}.day`, { validate: true }) }
+            onBlur={ validateAllHandler }
             fullWidth
           >
-            { _.map(days, (value, key) => <MenuItem key={ value + key } insetChildren checked={ availabilities[index] && availabilities[index].day === value } value={ value } primaryText={ <span> { value } </span> } />) }
+            { _.map(days, (value, index) => <MenuItem key={ value + index } insetChildren checked={ availabilities[index] && availabilities[index].day === value } value={ index } primaryText={ <span> { value } </span> } />) }
           </SelectField>
 
           <TimePicker
-            format="24hr"
+            format='24hr'
             hintText={
               <FormattedMessage
                 id='NewAvailability.selectStartTime'
                 defaultMessage='Start Time - 24Hr Format'
               />
             }
-            value={ availabilities[index]? availabilities[index].start_time: {} }
-            errorText={ _.get(errors, `${availability}.start_time`) }
-            onChange={ changeHandler(`${availability}.start_time`) }
+            value={ currentAvailability? start_time : { } }
+            errorText={ _.get(errors, `${index}.start_time`) }
+            onChange={ changeHandler(`${index}.start_time`, {  validate: true }) }
             onDismiss={ validateAllHandler }
             autoOk
             fullWidth
           />
 
           <TimePicker
-            format="24hr"
+            format='24hr'
             hintText={
               <FormattedMessage
                 id='NewAvailability.selectEndTime'
                 defaultMessage='End Time - 24Hr Format'
               />
             }
-            value={ availabilities[index]? availabilities[index].end_time: {} }
-            errorText={ _.get(errors, `${availability}.end_time`) }
-            onChange={ changeHandler(`${availability}.end_time`) }
+            value={ currentAvailability? end_time : { } }
+            errorText={ _.get(errors, `${index}.end_time`) }
+            onChange={ changeHandler(`${index}.end_time`, {  validate: true }) }
             onDismiss={ validateAllHandler }
             autoOk
             fullWidth
           />
+
           <FlatButton
             primary
             label={
@@ -288,7 +290,14 @@ class NewAvailability extends Component {
             }
             onClick={ this.handleAddAvailability }
           />
-          { index === 0 ? null : <FlatButton primary label={ <FormattedMessage id='NewAvailability.removeAvailability' defaultMessage='Remove availability' /> } onClick={ this.handleRemoveAvailability } /> }
+          { index === 0 ?
+            null :
+            <FlatButton
+              primary
+              label={ <FormattedMessage id='NewAvailability.removeAvailability' defaultMessage='Remove availability' /> }
+              onClick={ this.handleRemoveAvailability }
+            />
+          }
         </div>
       );
     });
@@ -326,7 +335,6 @@ class NewAvailability extends Component {
 
 
 NewAvailability.propTypes = {
-  match: PropTypes.object,
   errors: PropTypes.object,
   currentUser: PropTypes.shape({
     email: PropTypes.string,
@@ -337,25 +345,26 @@ NewAvailability.propTypes = {
     0: PropTypes.shape({
       start_time: PropTypes.object,
       end_time: PropTypes.object,
-      day: PropTypes.string,
+      day: PropTypes.number,
     })
   }),
   days: PropTypes.array,
   changeHandler: PropTypes.func.isRequired,
   changeValue: PropTypes.func.isRequired,
   validateAllHandler: PropTypes.func.isRequired,
-  validateHandler: PropTypes.func.isRequired,
+  validateAll: PropTypes.func.isRequired,
+  location: PropTypes.shape({
+    state: PropTypes.object
+  })
 };
 
 NewAvailability.defaultProps = {
-  match: {
-    params: {
-
-    }
-  },
   currentUser: {
     email: '',
     timezone: 'UTC',
+  },
+  location: {
+    state: {}
   },
   errors: {},
   days: [],
@@ -363,7 +372,7 @@ NewAvailability.defaultProps = {
     0: {
       start_time: {},
       end_time: {},
-      day: '',
+      day: null,
     }
   },
 };
