@@ -11,10 +11,12 @@ import FlatButton from 'material-ui/FlatButton';
 import TextField from 'material-ui/TextField';
 
 import FormData from './utils/FormData';
-import Header from './Header';
+import Header from './reusable/Header';
 import SnackBarComponent from './reusable/SnackBarComponent';
-import { postData } from './sendData';
-import Footer from './Footer';
+import { postData } from './utils/sendData';
+import Footer from './reusable/Footer';
+import formatLink from './utils/Link';
+import PageHeader from './reusable/PageHeader';
 
 const schema = {
   email: Joi.string().email({ minDomainAtoms: 2 }).required().options({
@@ -42,12 +44,19 @@ class NewPasswordPage extends Component {
     };
   }
   render() {
-    const { errors, changeHandler, validateHandler, currentUser: { email } } = this.props;
+    const { errors, changeHandler, validateHandler, currentUser: { email, locale } } = this.props;
 
     return (
       <div>
         <Header  />
-
+        <PageHeader
+          title={
+            <FormattedMessage
+              id='NewPasswordPage.header'
+              defaultMessage='New Password'
+            />
+          }
+        />
         <form className="signInContainer">
           <TextField
             name='email'
@@ -65,11 +74,11 @@ class NewPasswordPage extends Component {
           { this.renderSubmitButton() }
 
           <div className='signInLinkSecondaryContainer'>
-            <a href='/sign_up/client' className='signInLinkSecondary'>
+            <a href={ formatLink('/sign_up/client', locale) } className='signInLinkSecondary'>
               <FlatButton primary label={ <FormattedMessage id='signUpClient' /> } />
             </a>
 
-            <a href='/sign_up/volunteer' className='signInLinkSecondary'>
+            <a href={ formatLink('/sign_up/volunteer', locale) } className='signInLinkSecondary'>
               <FlatButton primary label={ <FormattedMessage id='signUpVolunteer' /> } />
             </a>
           </div>
