@@ -1,24 +1,52 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import Paper from 'material-ui/Paper';
+
 import { FormattedMessage } from 'react-intl';
+import RaisedButton from 'material-ui/RaisedButton';
+import Header from './reusable/Header';
+import Footer from './reusable/Footer';
+import './utils/CheckMobile';
+import Logo from './reusable/Logo';
+
+import './NotFoundPage.css';
+import formatLink from './utils/Link';
+import PageHeader from './reusable/PageHeader';
 
 class NotFoundPage extends Component {
-
   render() {
+    const { currentUser: { locale }, currentUser } = this.props;
     return (
-      <div>
-        <FormattedMessage
-          id='NotFoundPage.title'
-          defaultMessage='Oops!'
-        />
-        <FormattedMessage
-          id='NotFoundPage.notFoundPage'
-          defaultMessage='The page doesn’t exist or some other error occurred'
-        />
-        <FormattedMessage
-          id='NotFoundPage.linkToHome'
-          defaultMessage='Go to our'
-        />
+      <div className='notFoundPageContainer'>
+        <Header currentUser={ currentUser } />
+        <Paper zDepth={ 1 } className='paperOverride' rounded={ false }>
+          <Logo />
+          <PageHeader title={
+            <FormattedMessage
+              id='NotFoundPage.title'
+              defaultMessage='Oops!'
+            />
+            }
+          />
+
+          <div className='notFoundPageSection'>
+            <FormattedMessage
+              id='NotFoundPage.notFoundPage'
+              defaultMessage='The page doesn’t exist or some other error occurred.'
+            />
+          </div>
+          <a href={ formatLink('/', locale) }>
+            <RaisedButton
+              primary
+              label={
+                <FormattedMessage
+                  id='homeLink'
+                />
+              }
+            />
+          </a>
+        </Paper>
+        <Footer />
       </div>
     );
   }
@@ -26,11 +54,13 @@ class NotFoundPage extends Component {
 }
 
 NotFoundPage.propTypes = {
-
+  currentUser: PropTypes.object
 };
 
 NotFoundPage.defaultProps = {
-
+ currentUser: {
+   locale: ''
+ }
 };
 
 export default NotFoundPage;
