@@ -12,12 +12,14 @@ import TextField from 'material-ui/TextField';
 import Checkbox from 'material-ui/Checkbox';
 
 import FormData from './utils/FormData';
-import Header from './Header';
+import Header from './reusable/Header';
 import SnackBarComponent from './reusable/SnackBarComponent';
-import { postData } from './sendData';
+import { postData } from './utils/sendData';
+import formatLink from './utils/Link';
 
 import './SignIn.css';
-import Footer from './Footer';
+import Footer from './reusable/Footer';
+import PageHeader from './reusable/PageHeader';
 
 const schema = {
   email: Joi.string().email({ minDomainAtoms: 2 }).required().options({
@@ -27,7 +29,7 @@ const schema = {
     }
   }
 }),
-  password: Joi.string().min(8).required(),
+  password: Joi.string().required(),
   remember_me: Joi.boolean()
 };
 
@@ -37,6 +39,7 @@ class SignIn extends Component {
 
     this.handleSignIn = this.handleSignIn.bind(this);
     this.handleHideSnackBar = this.handleHideSnackBar.bind(this);
+    this.handleKeyDown = this.handleKeyDown.bind(this);
 
     this.state = {
       showSnackBar: false,
@@ -49,15 +52,22 @@ class SignIn extends Component {
       currentUser: {
         password,
         email,
-        remember_me
+        remember_me,
+        locale
       },
     } = this.props;
 
     return (
       <div>
         <Header  />
-
-        <form className="signInContainer">
+        <PageHeader title={
+          <FormattedMessage
+            id='signIn'
+            defaultMessage='Sign In'
+           />
+         }
+         />
+        <form className='signInContainer' onSubmit={ this.handleKeyDown }>
           <TextField
             name='email'
             value={ email }
@@ -90,10 +100,20 @@ class SignIn extends Component {
             label='Remember me'
           />
 
-          <RaisedButton primary label='Sign In' onClick={ this.handleSignIn } className='signInLink' />
+          <RaisedButton
+            primary
+            label={
+              <FormattedMessage
+                id='signIn'
+                defaultMessage='Sign In'
+              />
+            }
+            onClick={ this.handleSignIn }
+            className='signInLink'
+          />
 
           <div className='signInLinkSecondaryContainer'>
-            <a href='/password/new' className='signInLinkSecondary'>
+            <a href={ formatLink('/password/new', locale) } className='signInLinkSecondary'>
               <FlatButton
                 primary
                 label={
@@ -106,7 +126,7 @@ class SignIn extends Component {
               />
             </a>
 
-            <a href='/sign_up/client' className='signInLinkSecondary'>
+            <a href={ formatLink('/sign_up/client', locale) } className='signInLinkSecondary'>
               <FlatButton
                 primary
                 label={
@@ -117,7 +137,7 @@ class SignIn extends Component {
               />
             </a>
 
-            <a href='/sign_up/volunteer' className='signInLinkSecondary'>
+            <a href={ formatLink('/sign_up/volunteer', locale) } className='signInLinkSecondary'>
               <FlatButton
                 primary
                 label={
@@ -135,6 +155,12 @@ class SignIn extends Component {
         <Footer />
       </div>
     );
+  }
+
+  handleKeyDown(event) {
+    if (event.keycode === 13) {
+      this.handleSignIn();
+    }
   }
 
   renderSnackBar() {
@@ -155,8 +181,8 @@ class SignIn extends Component {
         url: '/sign_in',
         attributes,
         method: 'POST',
-        successCallBack: () => {
-          location.assign('/');
+        successCallBack: ({ currentUser: { locale } }) => {
+          location.assign(formatLink('/', locale));
         },
         errorCallBack: (message) => {
           this.setState({
