@@ -2,10 +2,14 @@ const LOWEST = 'lowest';
 const HIGHEST = 'highest';
 const OLDEST = 'first';
 const RECENT = 'last';
+const CLOSEST = 'closest';
+const NEWEST = 'newest';
 
 export {
   LOWEST,
   HIGHEST,
   OLDEST,
-  RECENT
+  RECENT,
+  CLOSEST,
+  NEWEST
 };
