@@ -8,7 +8,9 @@ const SignUpSchema = {
       }
     }
   }),
+  email_notification: Joi.bool(),
   role: Joi.string(),
+  locale: Joi.string(),
   programs: Joi.array().min(1).options({
     language: {
       array: {
@@ -25,7 +27,7 @@ const SignUpSchema = {
     }
   }),
 
-  first_name: Joi.string().required().regex(/^[a-zA-Z\u00C0-\u017F]+$/).options({
+  first_name: Joi.string().required().regex(/^[ a-zA-Z\u00C0-\u017F]+$/).options({
     language: {
       any: {
         required: 'Please enter your first name',
@@ -33,7 +35,8 @@ const SignUpSchema = {
       },
       string: {
         regex: {
-          base: 'Please enter a first name without invalid characters'
+          base: 'Please enter a first name without invalid characters',
+          name: 'Please enter a first name without invalid characters',
         }
       }
     }
@@ -47,16 +50,7 @@ const SignUpSchema = {
     }
   }),
 
-  thumbnail_image: Joi.object().keys({
-    image: Joi.object(),
-    url: Joi.string()
-  }).or('image', 'url').options({
-    language: {
-      object: {
-        missing: 'Must provide a thumbnail'
-      }
-    }
-  }),
+  thumbnail_image: Joi.alternatives().try(Joi.object(), Joi.string()),
 
   address: Joi.string().allow(''),
   city: Joi.string().allow(''),
@@ -89,7 +83,8 @@ const SignUpSchema = {
         allowOnly: 'Please agree to our terms of use'
       }
     }
-  })
+  }),
+  current_password: Joi.string().min(8).allow('')
 };
 
 export default SignUpSchema;
