@@ -1,11 +1,11 @@
 import React from 'react';
 import Joi from 'joi-browser';
 import ReactJoiValidation from 'react-joi-validation';
-import _ from 'lodash';
 import {
   BrowserRouter as Router,
   Route,
-  Switch
+  Switch,
+  Redirect
 } from 'react-router-dom';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
@@ -36,54 +36,106 @@ import TermsAndConditions from '../components/TermsAndConditions';
 import NewAvailability from '../components/NewAvailability';
 import AvailabilityIndexPage from '../components/AvailabilityIndexPage';
 import UserProfile from '../components/UserProfile';
+import SearchResults from '../components/SearchResults';
+import AboutPage from '../components/AboutPage';
+import ConversationPage from '../components/ConversationPage';
+import UserReviewIndexPage from '../components/UserReviewIndexPage';
+
+import ChooseLanguage from '../components/utils/ChooseLanguage';
+
 import './application.css';
 
 import localeData from '../../../build/locales/data.json';
-import SearchResults from '../components/SearchResults';
 
-addLocaleData([...en, ...es ]);
+import { ENGLISH, SPANISH } from '../components/utils/availableLocales';
 
-const language = _.includes(_.split(window.location.href), '=')? _.last(_.split(window.location.href, '=')) : 'en';
-
+addLocaleData([ ...en, ...es ]);
+const language = ChooseLanguage();
 const messages = localeData[language];
 
 const { render }  = ReactDom;
 
 ReactJoiValidation.setJoi(Joi);
 
-document.addEventListener('turbolinks:load', () => {
+function renderUserProfile(data, props) {
+  const isUserSame = function(){
+    if (data && data.currentUser && data && data.user && data.user) {
+      return data.user.url_slug === data.currentUser.url_slug;
+    }
+  }();
+
+  if (isUserSame) {
+    return <Redirect to='/' />;
+  } else {
+    return <UserProfile { ...data } { ...props } />;
+  }
+}
+
+document.addEventListener('DOMContentLoaded', () => {
   const container = document.getElementById('main');
   const node = document.getElementById('data');
 
   const data = node? JSON.parse(node.getAttribute('data')): {};
+  const locale = `(${ENGLISH}|${SPANISH})`;
 
   render(
     <IntlProvider locale={ language } messages={ messages }>
       <MuiThemeProvider muiTheme={ getMuiTheme(MuiTheme) }>
         <Router >
-          <div>
+          <Switch>
+
+            <Route exact path={ `/${locale}` } render={ () => <Homepage { ...data } /> } />
+
+            <Route exact path={ `/${locale}/inbox` } render={ (props) => <ConversationIndexPage { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/inbox/:id` } render={ (props) => <ConversationPage { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/messages/new` } render={ (props) => <NewMessagePage { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/search` } render={ (props) => <SearchBar { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/volunteers` } render={ (props) => <SearchResults { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/my_profile` } render={ (props) => <MyProfile { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/:user_id/(:order)` } render={ (props) => <UserReviewIndexPage { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/reviews/author/:url_slug` } render={ (props) => <UserReviewIndexPage { ...data } { ...props } /> } />
+
+            <Route
+              exact
+              path={ `/${locale}/profiles/:url_slug` }
+              render={ (props) => (renderUserProfile (data, props)) }
+            />
+            <Route exact path={ `/${locale}/password/new` } component={ NewPasswordPage } />
+            <Route exact path={ `/${locale}/password/edit` } render={ (props) => <ResetPasswordPage { ...props } /> } />
+            <Route exact path={ `/${locale}/availabilities/new` } render={ (props) => <NewAvailability { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/availabilities` } render={ (props) => <AvailabilityIndexPage { ...data }  { ...props } /> } />
+
+            <Route exact path={ `/${locale}/about` } render={ (props) => <AboutPage { ...data } { ...props } /> } />
+            <Route exact path={ `/${locale}/sign_in` } component={ SignIn } />
+            <Route exact path={ `/${locale}/sign_up/:role` } render={ (props) => <SignUp { ...data }  { ...props } /> } />
+            <Route exact path={ `/${locale}/terms_of_use` } render={ (props) => <TermsAndConditions { ...data } { ...props } /> } />
+
             <Route exact path='/' render={ () => <Homepage { ...data } /> } />
             <Route exact path='/inbox' render={ (props) => <ConversationIndexPage { ...data } { ...props } /> } />
+            <Route exact path='/inbox/:id' render={ (props) => <ConversationPage { ...data } { ...props } /> } />
             <Route exact path='/messages/new' render={ (props) => <NewMessagePage { ...data } { ...props } /> } />
-            <Switch>
-              <Route path='/search/:sign_up' render={ (props) => <SearchBar { ...data } { ...props } /> } />
-              <Route exact path='/search' render={ (props) => <SearchBar { ...data } { ...props } /> } />
-            </Switch>
-            <Route exact path='/available_volunteers' render={ (props) => <SearchResults { ...data } { ...props } /> } />
+            <Route exact path='/search' render={ (props) => <SearchBar { ...data } { ...props } /> } />
+            <Route exact path='/volunteers' render={ (props) => <SearchResults { ...data } { ...props } /> } />
             <Route exact path='/my_profile' render={ (props) => <MyProfile { ...data } { ...props } /> } />
-            <Route exact path='/profiles/:url_slug' render={ (props) => <UserProfile { ...data } { ...props } /> } />
-            <Route exact path='/sign_up/:role' render={ (props) => <SignUp { ...data }  { ...props } /> } />
-            <Route exact path='/sign_in' component={ SignIn } />
+            <Route exact path='/reviews/author/:url_slug' render={ (props) => <UserReviewIndexPage { ...data } { ...props } /> } />
+
+            <Route
+              exact
+              path='/profiles/:url_slug'
+              render={ (props) => (renderUserProfile (data, props)) }
+            />
             <Route exact path='/password/new' component={ NewPasswordPage } />
-            <Route path='/password/edit' render={ (props) => <ResetPasswordPage { ...props } /> } />
-            <Route exact path='/terms_of_use' component={ TermsAndConditions } />
-            <Switch>
-              <Route exact path='/availabilities/new/:sign_up' render={ (props) => <NewAvailability { ...data } { ...props } /> } />
-              <Route exact path='/availabilities/new' render={ (props) => <NewAvailability { ...data } { ...props } /> } />
-            </Switch>
+            <Route exact path='/password/edit' render={ (props) => <ResetPasswordPage { ...props } /> } />
+            <Route exact path='/availabilities/new' render={ (props) => <NewAvailability { ...data } { ...props } /> } />
             <Route exact path='/availabilities' render={ (props) => <AvailabilityIndexPage { ...data }  { ...props } /> } />
-            <Route path='/page_not_found' component={ NotFoundPage } />
-          </div>
+
+            <Route exact path='/about' render={ (props) => <AboutPage { ...data } { ...props } /> } />
+            <Route exact path='/sign_in' component={ SignIn } />
+            <Route exact path='/sign_up/:role' render={ (props) => <SignUp { ...data }  { ...props } /> } />
+            <Route exact path='/terms_of_use' render={ (props) => <TermsAndConditions { ...data } { ...props } /> } />
+
+            <Route path='/*' render={ () => <NotFoundPage { ...data } /> } />
+          </Switch>
         </Router>
 
       </MuiThemeProvider>
