@@ -3,28 +3,21 @@ module Contexts
     class Search
 
       def initialize(search_params, current_user)
-        @params, @user, @timezone = search_params, current_user, current_user[:timezone]
+        @params, @current_user, @timezone = search_params, current_user, current_user[:timezone]
       end
 
       def execute
         unless @params[:program].present?
-          raise Contexts::Availabilities::Errors::ProgramMissing, 'Program field is required.'
+          message = I18n.t('custom_errors.messages.missing_program')
+          raise Contexts::Availabilities::Errors::ProgramMissing, message
         end
 
         unless @params[:day].present?
-          raise Contexts::Availabilities::Errors::DayMissing, 'Day field is required.'
+          message = I18n.t('custom_errors.messages.missing_day')
+          raise Contexts::Availabilities::Errors::DayMissing, message
         end
 
-        if @params[:distance].present?
-          @existing_volunteers = User.search(@params, @timezone)
-                                   .near(@user.full_address, @params[:distance], :order => false)
-                                   .order(last_sign_in_at: :desc)
-
-        else
-          @existing_volunteers = User.search(@params, @timezone)
-                                   .order(last_sign_in_at: :desc)
-        end
-
+        @existing_volunteers = User.search(@params, @timezone, @current_user.full_address)
       end
     end
   end
