@@ -36,7 +36,12 @@ Rails.application.configure do
   config.action_mailer.raise_delivery_errors = true
   config.action_mailer.delivery_method = :letter_opener
   config.action_mailer.default_url_options = { :host => 'http://localhost:5000' }
-  config.mailer_sender = 'test@test.com'
+
+  config.roadie_options = {
+      protocol: 'http',
+      host: 'localhost',
+      port: '5000'
+  }
 
   config.action_mailer.perform_caching = false
 
