@@ -1,6 +1,5 @@
-# config/unicorn.rb
-worker_processes Integer(ENV["WEB_CONCURRENCY"] || 1)
-timeout 30
+worker_processes Integer(ENV["WEB_CONCURRENCY"] || 3)
+timeout 15
 preload_app true
 
 before_fork do |server, worker|
