@@ -28,9 +28,7 @@ Role.create({
 })
 
 TermsAndConditions.create({
-    description: 'some text',
-    version: 1,
-    id: 1
+    description: 'some text'
 })
 
 Program.create({
@@ -57,3 +55,10 @@ Language.create({
     name: 'Español',
     url_slug: 'espanol'
 })
+
+user = User.create!(email: 'admin@example.com', password: 'password', password_confirmation: 'password') if Rails.env.development?
+
+if user.present?
+  user.roles = Role.find_by_url_slug('admin')
+  user.save!
+end
