@@ -12,8 +12,8 @@ RSpec.describe Availability, type: :model do
 
   it { is_expected.to have_db_column(:created_at).of_type(:datetime) }
   it { is_expected.to have_db_column(:updated_at).of_type(:datetime) }
-  it { is_expected.to have_db_column(:start_time).of_type(:time) }
-  it { is_expected.to have_db_column(:end_time).of_type(:time) }
+  it { is_expected.to have_db_column(:start_time).of_type(:datetime) }
+  it { is_expected.to have_db_column(:end_time).of_type(:datetime) }
   it { is_expected.to have_db_column(:day).of_type(:string) }
   it { is_expected.to have_db_column(:vacation).of_type(:boolean) }
 end
